+++ tools/lfs.lua
@@ -1,3 +1,5 @@
+---@diagnostic disable: undefined-global,different-requires
+
 local ok, compat = pcall(require, "tools.lfs_compat")
 if ok then
   return compat